[Fix] 본문 입력 부분 최대 입력 크기 설정
본문에 입력할 수 있는 길이의 제한이 없어 DB에 설정된 길이보다 길
경우 에러가 발생하기때문에 textarea에 최대 입력 크기를 설정.

closed #295
