+++ web/src/containers/NewPostPage/styles.js
@@ -1,7 +1,7 @@
 import NewPostWrapper from './NewPostWrapper';
-import Input from './Input';
+import Content from './Content';
 import FileSelectLabel from './FileSelectLabel';
 import FileInput from './FileInput';
 import FileNameInput from './FileNameInput';
 
-export { NewPostWrapper, Input, FileSelectLabel, FileInput, FileNameInput };
+export { NewPostWrapper, Content, FileSelectLabel, FileInput, FileNameInput };
